requirements.txt
pillow==12.1.1
PyMuPDF==1.27.2.2
